Add engine class and test
